requirements.txt
resend>=0.8.0
python-dotenv>=1.0.0
cryptography>=41.0.0
apscheduler>=3.10.0
